+++ cypress/.eslintrc.cjs
@@ -6,7 +6,10 @@ module.exports = {
   rules: {
     'no-unsanitized/property': 0,
     'no-unused-expressions': 0,
+<<<<<<< HEAD
     'import/extensions': 0,
+=======
+>>>>>>> upstream/main
     'jsdoc/require-jsdoc': 0,
   },
 }