+++ rollup.config.js
@@ -1,6 +1,9 @@
 import { readFileSync } from 'fs'
 import babel from '@rollup/plugin-babel'
+<<<<<<< HEAD
 // todo: use @rollup/plugin-terser: https://github.com/rollup/plugins/pull/1323
+=======
+>>>>>>> upstream/main
 import terser from '@rollup/plugin-terser'
 
 const packageJson = JSON.parse(readFileSync('package.json'))