+++ src/instanceMethods/close.js
@@ -8,10 +8,19 @@ import * as dom from '../utils/dom/index.js'
 import { undoIOSfix } from '../utils/iosFix.js'
 import { undoScrollbar } from '../utils/scrollbarFix.js'
 
+<<<<<<< HEAD
 /*
  * Instance method to close sweetAlert
  */
 
+=======
+/**
+ * @param {SweetAlert2} instance
+ * @param {HTMLElement} container
+ * @param {boolean} returnFocus
+ * @param {Function} didClose
+ */
+>>>>>>> upstream/main
 function removePopupAndResetState(instance, container, returnFocus, didClose) {
   if (dom.isToast()) {
     triggerDidCloseAndDispose(instance, didClose)
@@ -40,6 +49,12 @@ function removePopupAndResetState(instance, container, returnFocus, didClose) {
   removeBodyClasses()
 }
 
+<<<<<<< HEAD
+=======
+/**
+ * Remove SweetAlert2 classes from body
+ */
+>>>>>>> upstream/main
 function removeBodyClasses() {
   dom.removeClass(
     [document.documentElement, document.body],
@@ -47,6 +62,14 @@ function removeBodyClasses() {
   )
 }
 
+<<<<<<< HEAD
+=======
+/**
+ * Instance method to close sweetAlert
+ *
+ * @param {any} resolveValue
+ */
+>>>>>>> upstream/main
 export function close(resolveValue) {
   resolveValue = prepareResolveValue(resolveValue)
 
@@ -66,6 +89,12 @@ export function close(resolveValue) {
   }
 }
 
+<<<<<<< HEAD
+=======
+/**
+ * @returns {boolean}
+ */
+>>>>>>> upstream/main
 export function isAwaitingPromise() {
   return !!privateProps.awaitingPromise.get(this)
 }
@@ -94,6 +123,12 @@ const triggerClosePopup = (instance) => {
   return true
 }
 
+<<<<<<< HEAD
+=======
+/**
+ * @param {any} error
+ */
+>>>>>>> upstream/main
 export function rejectPromise(error) {
   const rejectPromise = privateMethods.swalPromiseReject.get(this)
   handleAwaitingPromise(this)
@@ -103,16 +138,35 @@ export function rejectPromise(error) {
   }
 }
 
+<<<<<<< HEAD
 export const handleAwaitingPromise = (instance) => {
+=======
+/**
+ * @param {SweetAlert2} instance
+ */
+export const handleAwaitingPromise = (instance) => {
+  // @ts-ignore
+>>>>>>> upstream/main
   if (instance.isAwaitingPromise()) {
     privateProps.awaitingPromise.delete(instance)
     // The instance might have been previously partly destroyed, we must resume the destroy process in this case #2335
     if (!privateProps.innerParams.get(instance)) {
+<<<<<<< HEAD
+=======
+      // @ts-ignore
+>>>>>>> upstream/main
       instance._destroy()
     }
   }
 }
 
+<<<<<<< HEAD
+=======
+/**
+ * @param {any} resolveValue
+ * @returns {SweetAlertResult}
+ */
+>>>>>>> upstream/main
 const prepareResolveValue = (resolveValue) => {
   // When user calls Swal.close()
   if (typeof resolveValue === 'undefined') {
@@ -133,6 +187,14 @@ const prepareResolveValue = (resolveValue) => {
   )
 }
 
+<<<<<<< HEAD
+=======
+/**
+ * @param {SweetAlert2} instance
+ * @param {HTMLElement} popup
+ * @param {SweetAlertOptions} innerParams
+ */
+>>>>>>> upstream/main
 const handlePopupAnimation = (instance, popup, innerParams) => {
   const container = dom.getContainer()
   // If animation is supported, animate
@@ -150,6 +212,16 @@ const handlePopupAnimation = (instance, popup, innerParams) => {
   }
 }
 
+<<<<<<< HEAD
+=======
+/**
+ * @param {SweetAlert2} instance
+ * @param {HTMLElement} popup
+ * @param {HTMLElement} container
+ * @param {boolean} returnFocus
+ * @param {Function} didClose
+ */
+>>>>>>> upstream/main
 const animatePopup = (instance, popup, container, returnFocus, didClose) => {
   globalState.swalCloseEventFinishedCallback = removePopupAndResetState.bind(
     null,
@@ -166,11 +238,25 @@ const animatePopup = (instance, popup, container, returnFocus, didClose) => {
   })
 }
 
+<<<<<<< HEAD
 const triggerDidCloseAndDispose = (instance, didClose) => {
   setTimeout(() => {
     if (typeof didClose === 'function') {
       didClose.bind(instance.params)()
     }
+=======
+/**
+ * @param {SweetAlert2} instance
+ * @param {Function} didClose
+ */
+const triggerDidCloseAndDispose = (instance, didClose) => {
+  setTimeout(() => {
+    if (typeof didClose === 'function') {
+      // @ts-ignore
+      didClose.bind(instance.params)()
+    }
+    // @ts-ignore
+>>>>>>> upstream/main
     instance._destroy()
   })
 }