+++ src/staticMethods/fire.js
@@ -1,3 +1,12 @@
+<<<<<<< HEAD
+=======
+/**
+ * Main method to create a new SweetAlert2 popup
+ *
+ * @param  {...SweetAlertOptions} args
+ * @returns {Promise<SweetAlertResult>}
+ */
+>>>>>>> upstream/main
 export function fire(...args) {
   const Swal = this // eslint-disable-line @typescript-eslint/no-this-alias
   return new Swal(...args)