+++ src/utils/dom/renderers/renderInput.js
@@ -60,9 +60,17 @@ const showInput = (params) => {
   dom.show(inputContainer)
 
   // input autofocus
+<<<<<<< HEAD
   setTimeout(() => {
     dom.focusInput(input)
   })
+=======
+  if (params.inputAutoFocus) {
+    setTimeout(() => {
+      dom.focusInput(input)
+    })
+  }
+>>>>>>> upstream/main
 }
 
 /**